pyproject.toml
[tool.poetry]
name = "imagen"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.109.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlalchemy = "^2.0.25"
psycopg2 = "^2.9.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pillow = "^10.2.0"
boto3 = "^1.34.38"
black = "^24.1.1"
python-multipart = "^0.0.7"
requests = "^2.31.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
